Update script for newer metpy version
I used an older metpy version which produces a white space left of the skewT diagram due to the rotation of the T axis. This problem is solved with a newer metpy version but I had to make a few changes in the way the numbers like CAPE and CIN are rounded for the inset to use the round function only on a float which didn't matter before.